feat: Add AnswerComment and QuestionComment entities
This commit adds the `AnswerComment` and `QuestionComment` entities to the `forum` domain in the `clean-ddd` project. These entities represent comments made on answers and questions, respectively. They have properties such as `authorId`, `content`, `createdAt`, and `updatedAt`. The entities also have getter and setter methods for accessing and modifying their properties. Additionally, the entities have a `create` static method for creating instances of the entities.